+++ src/Models/DominionTech.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace OpenDominion\Models;
+
+/**
+ * OpenDominion\Models\DominionTech
+ *
+ * @property int $dominion_id
+ * @property int $tech_id
+ * @property \Illuminate\Support\Carbon|null $created_at
+ * @property \Illuminate\Support\Carbon|null $updated_at
+ * @property-read \OpenDominion\Models\Dominion $dominion
+ * @property-read \OpenDominion\Models\Tech $tech
+ */
+class DominionTech extends AbstractModel
+{
+    protected $table = 'dominion_techs';
+
+    public function dominion()
+    {
+        return $this->belongsTo(Dominion::class, 'dominion_id');
+    }
+
+    public function tech()
+    {
+        return $this->belongsTo(Tech::class, 'tech_id');
+    }
+}