+++ src/Services/Dominion/Actions/InvadeActionService.php
@@ -13,6 +13,7 @@ use OpenDominion\Exceptions\GameException;
 use OpenDominion\Models\Dominion;
 use OpenDominion\Models\GameEvent;
 use OpenDominion\Models\Unit;
+use OpenDominion\Services\Dominion\GovernmentService;
 use OpenDominion\Services\Dominion\HistoryService;
 use OpenDominion\Services\Dominion\ProtectionService;
 use OpenDominion\Services\Dominion\QueueService;
@@ -22,11 +23,6 @@ use OpenDominion\Traits\DominionGuardsTrait;
 class InvadeActionService
 {
     use DominionGuardsTrait;
-
-    /**
-     * @var float Number of boats protected per dock
-     */
-    protected const BOATS_PROTECTED_PER_DOCK = 2.5;
 
     /**
      * @var float Base percentage of boats sunk
@@ -59,14 +55,9 @@ class InvadeActionService
     protected const OVERWHELMED_PERCENTAGE = 15.0;
 
     /**
-     * @var float Percentage of attacker prestige used to cap prestige gains (plus bonus)
-     */
-    protected const PRESTIGE_CAP_PERCENTAGE = 10.0;
-
-    /**
      * @var int Bonus prestige when invading successfully
      */
-    protected const PRESTIGE_CHANGE_ADD = 20;
+    protected const PRESTIGE_CHANGE_ADD = 30;
 
     /**
      * @var float Base prestige % change for both parties when invading
@@ -78,6 +69,9 @@ class InvadeActionService
 
     /** @var CasualtiesCalculator */
     protected $casualtiesCalculator;
+
+    /** @var GovernmentService */
+    protected $governmentService;
 
     /** @var LandCalculator */
     protected $landCalculator;
@@ -130,6 +124,7 @@ class InvadeActionService
      *
      * @param BuildingCalculator $buildingCalculator
      * @param CasualtiesCalculator $casualtiesCalculator
+     * @param GovernmentService $governmentService
      * @param LandCalculator $landCalculator
      * @param MilitaryCalculator $militaryCalculator
      * @param NotificationService $notificationService
@@ -141,6 +136,7 @@ class InvadeActionService
     public function __construct(
         BuildingCalculator $buildingCalculator,
         CasualtiesCalculator $casualtiesCalculator,
+        GovernmentService $governmentService,
         LandCalculator $landCalculator,
         MilitaryCalculator $militaryCalculator,
         NotificationService $notificationService,
@@ -151,6 +147,7 @@ class InvadeActionService
     ) {
         $this->buildingCalculator = $buildingCalculator;
         $this->casualtiesCalculator = $casualtiesCalculator;
+        $this->governmentService = $governmentService;
         $this->landCalculator = $landCalculator;
         $this->militaryCalculator = $militaryCalculator;
         $this->notificationService = $notificationService;
@@ -259,6 +256,7 @@ class InvadeActionService
 
             $this->handleMoraleChanges($dominion, $target);
             $this->handleLandGrabs($dominion, $target);
+            $this->handleResearchPoints($dominion, $survivingUnits);
 
             $this->invasionResult['attacker']['unitsSent'] = $units;
 
@@ -366,13 +364,15 @@ class InvadeActionService
         if ($isOverwhelmed || ($range < 60)) {
             $attackerPrestigeChange = ($dominion->prestige * -(static::PRESTIGE_CHANGE_PERCENTAGE / 100));
         } elseif ($isInvasionSuccessful && ($range >= 75)) {
-            $attackerPrestigeChange = (int)round(min(
-                (($target->prestige * (($range / 100) / 10)) + static::PRESTIGE_CHANGE_ADD), // Gained through invading
-                (($dominion->prestige * (static::PRESTIGE_CAP_PERCENTAGE / 100)) + static::PRESTIGE_CHANGE_ADD) // But capped by depending on your current prestige
-            ));
-            $targetPrestigeChange = (int)round(($target->prestige * -(static::PRESTIGE_CHANGE_PERCENTAGE / 100)));
+            $attackerPrestigeChange = (int)round(($target->prestige * (($range / 100) / 10)) + static::PRESTIGE_CHANGE_ADD);
+            $targetPrestigeChange = (int)round($target->prestige * -(static::PRESTIGE_CHANGE_PERCENTAGE / 100));
 
-            // todo: if wat war, increase $attackerPrestigeChange by +15%
+            // War Bonus
+            if ($this->governmentService->isAtMutualWarWithRealm($dominion->realm, $target->realm)) {
+                $attackerPrestigeChange *= 1.25;
+            } elseif ($this->governmentService->isAtWarWithRealm($dominion->realm, $target->realm)) {
+                $attackerPrestigeChange *= 1.15;
+            }
         }
 
         // Reduce attacker prestige gain if the target was hit recently
@@ -393,6 +393,10 @@ class InvadeActionService
 
             $this->invasionResult['defender']['recentlyInvadedCount'] = $recentlyInvadedCount;
         }
+
+        // Cap for max prestige based on land size
+        $maxPrestigeChange = max(($this->landCalculator->getTotalLand($dominion) * 1.2) - $dominion->prestige, 0);
+        $attackerPrestigeChange = min($attackerPrestigeChange, $maxPrestigeChange);
 
         if ($attackerPrestigeChange !== 0) {
             if (!$isInvasionSuccessful) {
@@ -555,6 +559,10 @@ class InvadeActionService
         $targetDP = $this->invasionResult['defender']['dp'];
         $defensiveCasualtiesPercentage = (static::CASUALTIES_DEFENSIVE_BASE_PERCENTAGE / 100);
 
+        // Modify casualties percentage based on relative land size
+        $landRatio = $this->rangeCalculator->getDominionRange($dominion, $target) / 100;
+        $defensiveCasualtiesPercentage *= clamp($landRatio, 0.4, 1);
+
         // Scale casualties further with invading OP vs target DP
         $defensiveCasualtiesPercentage *= ($attackingForceOP / $targetDP);
 
@@ -652,9 +660,14 @@ class InvadeActionService
         $rangeMultiplier = ($range / 100);
 
         $landGrabRatio = 1;
-        // todo: if mutual war, $landGrabRatio = 1.2
-        // todo: if non-mutual war, $landGrabRatio = 1.15
         $bonusLandRatio = 1.7647;
+
+        // War Bonus
+        if ($this->governmentService->isAtMutualWarWithRealm($dominion->realm, $target->realm)) {
+            $landGrabRatio = 1.2;
+        } elseif ($this->governmentService->isAtWarWithRealm($dominion->realm, $target->realm)) {
+            $landGrabRatio = 1.15;
+        }
 
         $attackerLandWithRatioModifier = ($this->landCalculator->getTotalLand($dominion) * $landGrabRatio);
 
@@ -689,6 +702,9 @@ class InvadeActionService
 
             // Remove land
             $target->{"land_$landType"} -= $landLost;
+
+            // Add discounted land for buildings destroyed
+            $target->discounted_land += $buildingsToDestroy;
 
             // Destroy buildings
             foreach ($buildingsLostForLandType as $buildingType => $buildingsLost) {
@@ -889,6 +905,53 @@ class InvadeActionService
         return $convertedUnits;
     }
 
+    /**
+     * Handles research point generation for attacker.
+     *
+     * Original formula:
+     * (Conquered acres) * max([(Days into the round)/3], 10)
+     * - Past day 30 of the round, RP gains by attacking goes up and peaks at 16.6667 on day 50
+     * - This number is increased by 50% in the Base Ruleset.
+     *
+     * Removing days into round portion of the formula in favor of a static coefficient.
+     * For base ruleset, the range for the coefficent was 15-25. We are using 20 instead.
+     *
+     * @param Dominion $dominion
+     * @param array $units
+     */
+    protected function handleResearchPoints(Dominion $dominion, array $units): void
+    {
+        $researchPointsPerAcre = 20;
+
+        $isInvasionSuccessful = $this->invasionResult['result']['success'];
+        if ($isInvasionSuccessful) {
+            $landConquered = array_sum($this->invasionResult['attacker']['landConquered']);
+            $landGenerated = array_sum($this->invasionResult['attacker']['landGenerated']);
+
+            $researchPointsGained = ($landConquered + $landGenerated) * $researchPointsPerAcre;
+            $slowestTroopsReturnHours = $this->getSlowestUnitReturnHours($dominion, $units);
+
+            // Racial Bonus
+            $researchPointsGained *= (1 + $dominion->race->getPerkMultiplier('tech_production'));
+
+            $this->queueService->queueResources(
+                'invasion',
+                $dominion,
+                ['resource_tech' => $researchPointsGained],
+                $slowestTroopsReturnHours
+            );
+
+            $this->invasionResult['attacker']['researchPoints'] = $researchPointsGained;
+        }
+    }
+
+    /**
+     * Handles perks that trigger on invasion.
+     *
+     * @param Dominion $dominion
+     * @param Dominion $target
+     * @param array $units
+     */
     protected function handleAfterInvasionUnitPerks(Dominion $dominion, Dominion $target, array $units): void
     {
         // todo: just hobgoblin plunder atm, need a refactor later to take into
@@ -1011,7 +1074,7 @@ class InvadeActionService
             }
         }
         if (!$this->invasionResult['result']['overwhelmed'] && $unitsThatSinkBoats > 0) {
-            $defenderBoatsProtected = (static::BOATS_PROTECTED_PER_DOCK * $target->building_dock);
+            $defenderBoatsProtected = $this->militaryCalculator->getBoatsProtected($target);
             $defenderBoatsSunkPercentage = (static::BOATS_SUNK_BASE_PERCENTAGE / 100) * ($unitsThatSinkBoats / $unitsTotal);
             $targetQueuedBoats = $this->queueService->getInvasionQueueTotalByResource($target, 'resource_boats');
             $targetBoatTotal = $target->resource_boats + $targetQueuedBoats;