+++ src/Services/Dominion/Actions/SpellActionService.php
@@ -5,6 +5,7 @@ namespace OpenDominion\Services\Dominion\Actions;
 use DB;
 use Exception;
 use LogicException;
+use OpenDominion\Calculators\Dominion\ImprovementCalculator;
 use OpenDominion\Calculators\Dominion\LandCalculator;
 use OpenDominion\Calculators\Dominion\MilitaryCalculator;
 use OpenDominion\Calculators\Dominion\PopulationCalculator;
@@ -26,10 +27,8 @@ class SpellActionService
 {
     use DominionGuardsTrait;
 
-    /**
-     * @var float Info op base success rate
-     */
-    protected const INFO_MULTIPLIER_SUCCESS_RATE = 1.4;
+    /** @var ImprovementCalculator */
+    protected $improvementCalculator;
 
     /** @var LandCalculator */
     protected $landCalculator;
@@ -69,6 +68,7 @@ class SpellActionService
      */
     public function __construct()
     {
+        $this->improvementCalculator = app(ImprovementCalculator::class);
         $this->landCalculator = app(LandCalculator::class);
         $this->militaryCalculator = app(MilitaryCalculator::class);
         $this->networthCalculator = app(NetworthCalculator::class);
@@ -81,6 +81,8 @@ class SpellActionService
         $this->spellCalculator = app(SpellCalculator::class);
         $this->spellHelper = app(SpellHelper::class);
     }
+
+    public const BLACK_OPS_DAYS_AFTER_ROUND_START = 7;
 
     /**
      * Casts a magic spell for a dominion, optionally aimed at another dominion.
@@ -151,18 +153,9 @@ class SpellActionService
             } elseif ($this->spellHelper->isInfoOpSpell($spellKey)) {
                 $result = $this->castInfoOpSpell($dominion, $spellKey, $target);
 
-            } elseif ($this->spellHelper->isBlackOpSpell($spellKey)) {
-                if($dominion->round->hasOffensiveActionsDisabled())
-                {
-                    throw new GameException('Black ops have been disabled for the remainder of the round.');
-                }
-                throw new LogicException('Not yet implemented');
-            } elseif ($this->spellHelper->isWarSpell($spellKey)) {
-                if($dominion->round->hasOffensiveActionsDisabled())
-                {
-                    throw new GameException('Black ops have been disabled for the remainder of the round.');
-                }
-                throw new LogicException('Not yet implemented');
+            } elseif ($this->spellHelper->isHostileSpell($spellKey)) {
+                $result = $this->castHostileSpell($dominion, $spellKey, $target);
+
             } else {
                 throw new LogicException("Unknown type for spell {$spellKey}");
             }
@@ -288,8 +281,7 @@ class SpellActionService
 
         // 100% spell success if target has a WPA of 0
         if ($targetWpa !== 0.0) {
-            $successRate = $this->opsHelper->operationSuccessChance($selfWpa, $targetWpa,
-                static::INFO_MULTIPLIER_SUCCESS_RATE);
+            $successRate = $this->opsHelper->infoOperationSuccessChance($selfWpa, $targetWpa);
 
             if (!random_chance($successRate)) {
                 // Inform target that they repelled a hostile spell
@@ -297,6 +289,7 @@ class SpellActionService
                     ->queueNotification('repelled_hostile_spell', [
                         'sourceDominionId' => $dominion->id,
                         'spellKey' => $spellKey,
+                        'unitsKilled' => '',
                     ])
                     ->sendNotifications($target, 'irregular_dominion');
 
@@ -304,13 +297,11 @@ class SpellActionService
                 return [
                     'success' => false,
                     'message' => "The enemy wizards have repelled our {$spellInfo['name']} attempt.",
-                    'wizardStrengthCost' => 2,
+                    'wizardStrengthCost' => 1,
                     'alert-type' => 'warning',
                 ];
             }
         }
-
-        // todo: take Energy Mirror into account with 20% spell reflect (either show your info or give the infoop to the target)
 
         $infoOp = new InfoOp([
             'source_realm_id' => $dominion->realm->id,
@@ -356,9 +347,12 @@ class SpellActionService
                 ];
                 break;
 
-//            case 'vision':
-//                $infoOp->data = [];
-//                break;
+            case 'vision':
+                $infoOp->data = [
+                    'techs' => $target->techs->pluck('name', 'key')->all(),
+                    'heroes' => []
+                ];
+                break;
 
             case 'revelation':
                 $infoOp->data = $this->spellCalculator->getActiveSpells($target);
@@ -388,9 +382,327 @@ class SpellActionService
         return [
             'success' => true,
             'message' => 'Your wizards cast the spell successfully, and a wealth of information appears before you.',
-            'wizardStrengthCost' => 2,
+            'wizardStrengthCost' => 1,
             'redirect' => $redirect,
         ];
+    }
+
+    /**
+     * Casts a hostile spell for $dominion to $target.
+     *
+     * @param Dominion $dominion
+     * @param string $spellKey
+     * @param Dominion $target
+     * @return array
+     * @throws GameException
+     * @throws LogicException
+     */
+    protected function castHostileSpell(Dominion $dominion, string $spellKey, Dominion $target): array
+    {
+        if ($dominion->round->hasOffensiveActionsDisabled()) {
+            throw new GameException('Black ops have been disabled for the remainder of the round.');
+        }
+
+        if (now()->diffInDays($dominion->round->start_date) < self::BLACK_OPS_DAYS_AFTER_ROUND_START) {
+            throw new GameException('You cannot perform black ops for the first seven days of the round');
+        }
+
+        $spellInfo = $this->spellHelper->getSpellInfo($spellKey, $dominion->race);
+
+        if ($this->spellHelper->isWarSpell($spellKey)) {
+            $warDeclared = ($dominion->realm->war_realm_id == $target->realm->id || $target->realm->war_realm_id == $dominion->realm->id);
+            if (!$warDeclared && !$this->militaryCalculator->recentlyInvadedBy($dominion, $target)) {
+                throw new GameException("You cannot cast {$spellInfo['name']} outside of war.");
+            }
+        }
+
+        $selfWpa = $this->militaryCalculator->getWizardRatio($dominion, 'offense');
+        $targetWpa = $this->militaryCalculator->getWizardRatio($target, 'defense');
+
+        // You need at least some positive WPA to cast info ops
+        if ($selfWpa === 0.0) {
+            // Don't reduce mana by throwing an exception here
+            throw new GameException("Your wizard force is too weak to cast {$spellInfo['name']}. Please train more wizards.");
+        }
+
+        // 100% spell success if target has a WPA of 0
+        if ($targetWpa !== 0.0) {
+            $successRate = $this->opsHelper->blackOperationSuccessChance($selfWpa, $targetWpa);
+
+            if (!random_chance($successRate)) {
+                $wizardsKilledBasePercentage = 1;
+
+                $wizardLossSpaRatio = ($targetWpa / $selfWpa);
+                $wizardsKilledPercentage = clamp($wizardsKilledBasePercentage * $wizardLossSpaRatio, 0.5, 1.5);
+
+                $unitsKilled = [];
+                $wizardsKilled = (int)floor($dominion->military_wizards * ($wizardsKilledPercentage / 100));
+
+                // Check for immortal wizards
+                if ($dominion->race->getPerkValue('immortal_wizards') != 0) {
+                    $wizardsKilled = 0;
+                }
+
+                if ($wizardsKilled > 0) {
+                    $unitsKilled['wizards'] = $wizardsKilled;
+                    $dominion->military_wizards -= $wizardsKilled;
+                }
+
+                foreach ($dominion->race->units as $unit) {
+                    if ($unit->getPerkValue('counts_as_wizard_offense')) {
+                        $unitKilledMultiplier = ((float)$unit->getPerkValue('counts_as_wizard_offense') / 2) * ($wizardsKilledPercentage / 100);
+                        $unitKilled = (int)floor($dominion->{"military_unit{$unit->slot}"} * $unitKilledMultiplier);
+                        if ($unitKilled > 0) {
+                            $unitsKilled[strtolower($unit->name)] = $unitKilled;
+                            $dominion->{"military_unit{$unit->slot}"} -= $unitKilled;
+                        }
+                    }
+                }
+
+                $unitsKilledStringParts = [];
+                foreach ($unitsKilled as $name => $amount) {
+                    $amountLabel = number_format($amount);
+                    $unitLabel = str_plural(str_singular($name), $amount);
+                    $unitsKilledStringParts[] = "{$amountLabel} {$unitLabel}";
+                }
+                $unitsKilledString = generate_sentence_from_array($unitsKilledStringParts);
+
+                // Inform target that they repelled a hostile spell
+                $this->notificationService
+                    ->queueNotification('repelled_hostile_spell', [
+                        'sourceDominionId' => $dominion->id,
+                        'spellKey' => $spellKey,
+                        'unitsKilled' => $unitsKilledString,
+                    ])
+                    ->sendNotifications($target, 'irregular_dominion');
+
+                if ($unitsKilledString) {
+                    $message = "The enemy wizards have repelled our {$spellInfo['name']} attempt and managed to kill $unitsKilledString.";
+                } else {
+                    $message = "The enemy wizards have repelled our {$spellInfo['name']} attempt.";
+                }
+
+                // Return here, thus completing the spell cast and reducing the caster's mana
+                return [
+                    'success' => false,
+                    'message' => $message,
+                    'wizardStrengthCost' => 5,
+                    'alert-type' => 'warning',
+                ];
+            }
+        }
+
+        $spellReflected = false;
+        if ($this->spellCalculator->isSpellActive($target, 'energy_mirror') && random_chance(0.2)) {
+            $spellReflected = true;
+            $deflectedBy = $target;
+            $target = $dominion;
+            $dominion = $deflectedBy;
+        }
+
+        if (isset($spellInfo['duration'])) {
+            // Cast spell with duration
+            if ($this->spellCalculator->isSpellActive($target, $spellKey)) {
+                $where = [
+                    'dominion_id' => $target->id,
+                    'spell' => $spellKey,
+                ];
+
+                $activeSpell = DB::table('active_spells')
+                    ->where($where)
+                    ->first();
+
+                if ($activeSpell === null) {
+                    throw new LogicException("Active spell '{$spellKey}' for dominion id {$target->id} not found");
+                }
+
+                DB::table('active_spells')
+                    ->where($where)
+                    ->update([
+                        'duration' => $spellInfo['duration'],
+                        'cast_by_dominion_id' => $dominion->id,
+                        'updated_at' => now(),
+                    ]);
+            } else {
+                DB::table('active_spells')
+                    ->insert([
+                        'dominion_id' => $target->id,
+                        'spell' => $spellKey,
+                        'duration' => $spellInfo['duration'],
+                        'cast_by_dominion_id' => $dominion->id,
+                        'created_at' => now(),
+                        'updated_at' => now(),
+                    ]);
+            }
+
+            // Update statistics
+            if (isset($dominion->{"stat_{$spellInfo['key']}_hours"})) {
+                $dominion->{"stat_{$spellInfo['key']}_hours"} += $spellInfo['duration'];
+            }
+
+            // Surreal Perception
+            $sourceDominionId = null;
+            if ($this->spellCalculator->isSpellActive($target, 'surreal_perception')) {
+                $sourceDominionId = $dominion->id;
+            }
+
+            $this->notificationService
+                ->queueNotification('received_hostile_spell', [
+                    'sourceDominionId' => $sourceDominionId,
+                    'spellKey' => $spellKey,
+                ])
+                ->sendNotifications($target, 'irregular_dominion');
+
+            if ($spellReflected) {
+                // Notification for Energy Mirror deflection
+                $this->notificationService
+                    ->queueNotification('reflected_hostile_spell', [
+                        'sourceDominionId' => $target->id,
+                        'spellKey' => $spellKey,
+                    ])
+                    ->sendNotifications($dominion, 'irregular_dominion');
+
+                return [
+                    'success' => true,
+                    'message' => sprintf(
+                        'Your wizards cast the spell successfully, but it was deflected and it will now affect your dominion for the next %s hours.',
+                        $spellInfo['duration']
+                    ),
+                    'alert-type' => 'danger'
+                ];
+            } else {
+                return [
+                    'success' => true,
+                    'message' => sprintf(
+                        'Your wizards cast the spell successfully, and it will continue to affect your target for the next %s hours.',
+                        $spellInfo['duration']
+                    )
+                ];
+            }
+        } else {
+            // Cast spell instantly
+            $damageDealt = [];
+            $totalDamage = 0;
+            $baseDamage = (isset($spellInfo['percentage']) ? $spellInfo['percentage'] : 1) / 100;
+
+            if (isset($spellInfo['decreases'])) {
+                foreach ($spellInfo['decreases'] as $attr) {
+                    $damage = $target->{$attr} * $baseDamage;
+
+                    // Damage reduction from Forest Havens
+                    if ($attr == 'peasants') {
+                        $forestHavenFireballReduction = 8;
+                        $forestHavenFireballReductionMax = 80;
+                        $damageMultiplier = (1 - min(
+                            (($target->building_forest_haven / $this->landCalculator->getTotalLand($target)) * $forestHavenFireballReduction),
+                            ($forestHavenFireballReductionMax / 100)
+                        ));
+                        $damage *= $damageMultiplier;
+                    }
+
+                    // Damage reduction from Masonries
+                    if (strpos($attr, 'improvement_') === 0) {
+                        $masonryLightningBoltReduction = 0.75;
+                        $masonryLightningBoltReductionMax = 25;
+                        $damageMultiplier = (1 - min(
+                            (($target->building_forest_haven / $this->landCalculator->getTotalLand($target)) * $masonryLightningBoltReduction),
+                            ($masonryLightningBoltReductionMax / 100)
+                        ));
+                        $damage *= $damageMultiplier;
+                    }
+
+                    // Damage reduction from Towers
+                    $damage *= (1 - $this->improvementCalculator->getImprovementMultiplierBonus($target, 'towers'));
+
+                    $totalDamage += round($damage);
+                    $target->{$attr} -= round($damage);
+                    $damageDealt[] = sprintf('%s %s', number_format($damage), dominion_attr_display($attr, $damage));
+
+                    // Update statistics
+                    if (isset($dominion->{"stat_{$spellInfo['key']}_damage"})) {
+                        // Only count peasants killed by fireball
+                        if (!($spellInfo['key'] == 'fireball' && $attr == 'resource_food')) {
+                            $dominion->{"stat_{$spellInfo['key']}_damage"} += round($damage);
+                        }
+                    }
+                }
+
+                // Combine lightning bolt damage into single string
+                if ($spellInfo['key'] === 'lightning_bolt') {
+                    // Combine lightning bold damage into single string
+                    $damageDealt = [sprintf('%s %s', number_format($totalDamage), dominion_attr_display('improvement', $totalDamage))];
+                }
+            }
+            if (isset($spellInfo['increases'])) {
+                foreach ($spellInfo['increases'] as $attr) {
+                    $damage = $target->{$attr} * $baseDamage;
+
+                    // Damage reduction from Towers
+                    $damage *= (1 - $this->improvementCalculator->getImprovementMultiplierBonus($target, 'towers'));
+
+                    $target->{$attr} += round($damage);
+                }
+            }
+
+            $target->save([
+                'event' => HistoryService::EVENT_ACTION_CAST_SPELL,
+                'action' => $spellKey
+            ]);
+
+            // Prestige Gains
+            $prestigeGainString = '';
+            if ($this->spellHelper->isWarSpell($spellKey) && ($dominion->realm->war_realm_id == $target->realm->id && $target->realm->war_realm_id == $dominion->realm->id) && $totalDamage > 0) {
+                $dominion->prestige += 2;
+                $dominion->stat_wizard_prestige += 2;
+                $prestigeGainString = 'You were awarded 2 prestige due to mutual war.';
+            }
+
+            // Surreal Perception
+            $sourceDominionId = null;
+            if ($this->spellCalculator->isSpellActive($target, 'surreal_perception')) {
+                $sourceDominionId = $dominion->id;
+            }
+
+            $damageString = generate_sentence_from_array($damageDealt);
+
+            $this->notificationService
+                ->queueNotification('received_hostile_spell', [
+                    'sourceDominionId' => $sourceDominionId,
+                    'spellKey' => $spellKey,
+                    'damageString' => $damageString,
+                ])
+                ->sendNotifications($target, 'irregular_dominion');
+
+            if ($spellReflected) {
+                // Notification for Energy Mirror defelection
+                $this->notificationService
+                    ->queueNotification('reflected_hostile_spell', [
+                        'sourceDominionId' => $target->id,
+                        'spellKey' => $spellKey,
+                    ])
+                    ->sendNotifications($dominion, 'irregular_dominion');
+
+                return [
+                    'success' => true,
+                    'message' => sprintf(
+                        'Your wizards cast the spell successfully, but it was deflected and your dominion lost %s.',
+                        $damageString
+                    ),
+                    'wizardStrengthCost' => 5,
+                    'alert-type' => 'danger'
+                ];
+            } else {
+                return [
+                    'success' => true,
+                    'message' => sprintf(
+                        'Your wizards cast the spell successfully, your target lost %s. %s',
+                        $damageString,
+                        $prestigeGainString
+                    ),
+                    'wizardStrengthCost' => 5,
+                ];
+            }
+        }
     }
 
     /**